Remove redundant client lib and state directories
