--- client/state/index.ts
@@ -1,13 +1,0 @@
-import { createSlice } from "@reduxjs/toolkit";
-
-export const initialState = {};
-
-export const globalSlice = createSlice({
-  name: "global",
-  initialState,
-  reducers: {},
-});
-
-export const {} = globalSlice.actions;
-
-export default globalSlice.reducer;
